Update to null safety on dart from flutter channel beta

+++ lib/src/codec.dart
@@ -4,7 +4,8 @@ import 'dart:typed_data';
 
 class Codec {
   static int fold(Iterable<int> bytes) =>
-      bytes.reduce((x, y) => x * pow(2, 8) + y);
+      // ignore: avoid_as
+      bytes.reduce((x, y) => (x * pow(2, 8) as int) + y);
 
   static Iterable<int> unfold(int n) => [
         for (var i in [3, 2, 1, 0]) (n ~/ pow(256, i)) % 256
